Added http_client attribute to DIDWeb to avoid costly construction per resolution.
DIDWeb struct has an HTTP client to use for DID resolution.  It's incredibly slow to create a new
reqwest::Client, due to the overhead of loading the system's root certificates.  This HTTP client must
be specified by constructing the DIDWeb instance using new_with_default_http_client (to use defaults)
or DIDWeb::new_with_http_client if there is an specific reqwest::Client that should be reused.
Note that this is the recommended approach to using reqwest::Client (see
https://docs.rs/reqwest/latest/reqwest/struct.Client.html).