Change ssi-contexts license to pure Apache2 like other crates
After careful analysis, the CC-BY-SA license (as well as the others)
considers such library as a collection, and mostly require recognition
of the ownership of the context files, which we are already doing in the
README and LICENSE files. Therefore ssi-contexts doesn't need to use
those licenses itself, and we can have it be pure Apache2 like the rest
of the crates.